backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.60.0
python-dotenv>=1.0.0
